\name{fnoiser-package}
\alias{fnoiser-package}
\alias{fnoiser}
\docType{package}
\title{
1/f Noise generation routine for R
}
\description{
Generates one or more vectors of coloured noise.
}
\details{
\tabular{ll}{
Package: \tab fnoise\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-09-15\cr
License: \tab What license is it under?\cr
}
This package has only one function 'fnoise'.
}
\author{
Jani Anttila

Maintainer: Jani Anttila <jani.anttila@helsinki.fi>
}
\references{
Cohen et al.
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ noise }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
noise <- fn1(10,0.0)
}
