\name{fnoiser-package}
\alias{fnoiser-package}
\alias{fnoiser}
\docType{package}
\title{
  1/f Noise generation routine for R
}
\description{
  Generates coloured noise by summing sine waves.
}
\details{
  \tabular{ll}{
    Package: \tab fnoise\cr
    Type: \tab Package\cr
    Version: \tab 1.0\cr
    Date: \tab 2017-04-27\cr
    License: \tab What license is it under?\cr
  }
  This package has currently only one function 'fn1(t,g)'.
}
\author{
  Jani Anttila

  Maintainer: Jani Anttila <jani.anttila@helsinki.fi>
}
\references{
  Cohen et al. (1998) Proc. R. Soc. B.
}
\keyword{ noise }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
  noise <- fn1(10,0.0)
}
